# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clientes API",
        "description": "API de Gerenciamento da Clientes",
        "contact": {
            "email": "bsouza@fedora.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:83",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Ver um novo cliente",
                "operationId": "5f03d7ebe7fe53864c6c9f8a8a1d697e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente que será atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Zeca"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar um novo cliente",
                "operationId": "2268cf0da18a8f5cfffd5a4e86e8c90b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente que será atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Zeca"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar um novo cliente",
                "operationId": "3fa2e224a3463288f0ce9a0543210025",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente que será atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Zeca"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar e filtrar clientes",
                "operationId": "eb92d13d84d046f99e508a1ec80bf943",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClienteRequest"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar um novo cliente",
                "operationId": "4d55d7921508eac8e64e5ada3ba5f8c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Zeca"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ClienteRequestItem": {
                "type": "object"
            },
            "ClienteRequestDel": {
                "required": [
                    "nome",
                    "email"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Zeca"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "teste@teste.com"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "5598785",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ClienteRequest": {
                "required": [
                    "nome",
                    "email"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Zeca"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "teste@teste.com"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "5598785",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ClienteRequestUpd": {
                "required": [
                    "nome",
                    "email"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Zeca"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "teste@teste.com"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "5598785",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Clientes",
            "description": "Clientes"
        }
    ]
}